+++ packages/schema-to-zod/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm-schema-to-zod
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.10.2
+
 ## 0.3.1
 
 ### Patch Changes